# api/openapi.yaml
openapi: 3.1.0
info:
  title: meds-api
  version: "1.0.0"
  summary: API REST en lecture seule sur la Base de données publique des médicaments (BDPM).
  description: |
    API REST exposant la **Base de données publique des médicaments** (BDPM), publiée par l'ANSM
    avec le concours de la HAS et du ministère de la Santé.

    ## Avertissement

    Les données proviennent de la Base de données publique des médicaments (ANSM), diffusée sous
    Licence Ouverte (Etalab). Elles sont fournies **à titre informatif** et **ne se substituent ni
    au Résumé des Caractéristiques du Produit, ni à l'avis d'un professionnel de santé**. Cette API
    ne constitue pas un dispositif médical et n'apporte aucune aide à la décision clinique.
    La date de mise à jour du jeu de données est exposée sur `/v1/dataset`.

    ## Conventions

    - Toutes les routes `/v1/*` exigent une clé API (`Authorization: Bearer <clé>`).
    - Les montants sont des **entiers en centimes**, suffixés `_cents`.
    - L'absence de valeur est toujours `null`, jamais `""` ni `0`. La distinction compte :
      7 276 des 20 903 présentations n'ont aucun prix, et les confondre avec un prix nul
      fausserait toute statistique.
    - Les valeurs métier sont **conservées telles quelles** (`Autorisation active`,
      `Commercialisée`), sans traduction ni normalisation.
    - Les erreurs suivent la **RFC 9457** (`application/problem+json`).
  contact:
    name: Support meds-api
...
  license:
...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
paths:
  /v1/medicaments:
    get:
      tags: [Médicaments]
      summary: Recherche et filtres
      description: |
        Recherche plein-texte sur la dénomination, les substances et le titulaire, combinée à des
        filtres. **Les filtres se combinent en ET.** Sans `q`, le tri par pertinence bascule
        automatiquement sur la dénomination.
      operationId: listMedicaments
      parameters:
        - name: q
          in: query
          description: Recherche plein-texte. Tolère les accents et les fautes de frappe.
          schema: { type: string, maxLength: 200 }
          example: doliprane
        - name: commercialise
          in: query
          schema: { type: boolean }
        - name: statut_amm
          in: query
          schema:
            type: string
            enum: [active, abrogee, archivee, retiree, suspendue]
        - name: procedure_amm
          in: query
          schema:
            type: string
            enum:
              [
                nationale,
                centralisee,
                decentralisee,
                reconnaissance_mutuelle,
                homeo,
                phyto,
                importation_parallele,
              ]
        - name: forme
          in: query
          description: Forme pharmaceutique, correspondance exacte insensible à la casse.
          schema: { type: string }
          example: comprimé
        - name: voie
          in: query
          schema: { type: string }
          example: orale
        - name: titulaire
          in: query
          schema: { type: string }
        - name: substance
          in: query
          description: Code **ou** dénomination de substance, indifféremment.
          schema: { type: string }
          example: PARACETAMOL
        - name: surveillance
          in: query
          description: Surveillance renforcée (triangle noir).
          schema: { type: boolean }
        - name: mitm
          in: query
          description: Médicament d'intérêt thérapeutique majeur.
          schema: { type: boolean }
        - name: generique
          in: query
          schema:
            type: string
            enum: [princeps, generique, complementarite, substituable]
        - name: remboursable
          in: query
          schema: { type: boolean }
        - name: rupture
          in: query
          schema: { type: boolean }
        - name: sort
          in: query
          schema:
            type: string
            enum: [pertinence, denomination, date_amm]
            default: pertinence
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Liste paginée de spécialités.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/MedicamentListe" }
                  meta: { $ref: "#/components/schemas/MetaListe" }
              example:
                data:
                  - cis: "60234100"
                    denomination: "DOLIPRANE 1000 mg, comprimé"
                    forme_pharmaceutique: comprimé
                    voies_administration: [orale]
                    statut_amm: Autorisation active
                    etat_commercialisation: Commercialisée
                    date_amm: "1995-06-12"
                    titulaires: [OPELLA HEALTHCARE FRANCE]
                    surveillance_renforcee: false
                    substances: [PARACÉTAMOL]
                    nb_presentations: 3
                    mitm: false
                    score: 70.3
                    _links:
                      self: /v1/medicaments/60234100
                      presentations: /v1/medicaments/60234100/presentations
                meta:
                  total: 27
                  limit: 20
                  has_more: true
                  next_cursor: eyJvIjoyMCwiaCI6IjNmMmE5YyJ9
                  dataset_version: "2026-08-15T04:12:33Z"
        "304": { $ref: "#/components/responses/NotModified" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "406": { $ref: "#/components/responses/NotAcceptable" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/NotReady" }
...
  /v1/presentations/{cip}:
    get:
      tags: [Présentations]
      summary: Lookup par code-barres
      description: |
        Accepte indifféremment un **CIP7** (7 chiffres) ou un **CIP13** (13 chiffres) ; la
        distinction se fait sur la longueur. C'est l'endpoint du scan en officine.

        Le médicament parent est **inclus d'office** : après un scan, l'appelant veut toujours
        savoir de quel médicament il s'agit, et lui imposer un second appel serait un défaut de
        conception.
      operationId: getPresentation
      parameters:
        - name: cip
          in: path
          required: true
          schema: { type: string, pattern: "^([0-9]{7}|[0-9]{13})$" }
          example: "3400935955838"
      responses:
        "200":
          description: Présentation et son médicament parent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/PresentationDetail" }
                  meta: { $ref: "#/components/schemas/MetaSimple" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/substances:
    get:
      tags: [Référentiels]
      summary: Référentiel des substances
      description: Référentiel des substances actives, construit par agrégation de la composition des spécialités. `q` filtre sur la dénomination ou le code.
      operationId: listSubstances
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Liste paginée de substances.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Substance" } }
                  meta: { $ref: "#/components/schemas/MetaListe" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/substances/{code}:
    get:
      tags: [Référentiels]
      summary: Une substance
      description: Une substance et le nombre de spécialités qui la contiennent.
      operationId: getSubstance
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
          example: "42215"
      responses:
        "200":
          description: Substance et nombre de spécialités la contenant.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Substance" }
                  meta: { $ref: "#/components/schemas/MetaSimple" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/substances/{code}/medicaments:
    get:
      tags: [Référentiels]
      summary: Spécialités contenant une substance
      description: Spécialités contenant cette substance, paginées.
      operationId: getSubstanceMedicaments
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Liste paginée de spécialités.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/MedicamentListe" } }
                  meta: { $ref: "#/components/schemas/MetaListe" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/groupes-generiques:
    get:
      tags: [Référentiels]
      summary: Référentiel des groupes génériques
      description: |
        Seuls les groupes comptant **au moins un membre dans le référentiel courant** sont servis :
        1 524 sur les 1 671 présents dans la source, les 147 autres ne contenant que des
        spécialités retirées du marché.
      operationId: listGroupes
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Liste paginée de groupes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/GroupeGenerique" } }
                  meta: { $ref: "#/components/schemas/MetaListe" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/groupes-generiques/{id}:
    get:
      tags: [Référentiels]
      summary: Un groupe générique et ses membres
      description: Un groupe générique et tous ses membres, avec leur type (princeps, générique) et leur état de commercialisation.
      operationId: getGroupe
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: "1"
      responses:
        "200":
          description: Groupe et membres, avec leur type et leur état de commercialisation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/GroupeGenerique" }
                  meta: { $ref: "#/components/schemas/MetaSimple" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/ruptures:
    get:
      tags: [Référentiels]
      summary: Disponibilité des médicaments
      description: |
        `actives` vaut `true` par défaut : la question posée à cet endpoint est « qu'est-ce qui
        manque en ce moment ? », pas « qu'est-ce qui a manqué un jour ». Le défaut inverse noierait
        le signal utile.
      operationId: listRuptures
      parameters:
        - name: statut
          in: query
          schema:
            type: string
            enum: [rupture, tension, arret, remise_disposition]
        - name: actives
          in: query
          description: Exclut les remises à disposition passées.
          schema: { type: boolean, default: true }
        - name: cis
          in: query
          schema: { type: string, pattern: "^[0-9]{8}$" }
        - name: depuis
          in: query
          description: Fiches mises à jour depuis cette date.
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Liste paginée de fiches de disponibilité.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Rupture" } }
                  meta: { $ref: "#/components/schemas/MetaListe" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/mitm:
    get:
      tags: [Référentiels]
      summary: Médicaments d'intérêt thérapeutique majeur
      description: Médicaments d'intérêt thérapeutique majeur, au sens de l'arrêté du ministère de la Santé.
      operationId: listMITM
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Liste paginée de MITM.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/MITMListe" } }
                  meta: { $ref: "#/components/schemas/MetaListe" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/suggest:
    get:
      tags: [Référentiels]
      summary: Autocomplétion
      description: |
        Optimisé pour la frappe au kilomètre : préfixes uniquement, charge utile minimale.
        Le champ `highlight` est **échappé en HTML** avant insertion des balises `<em>` ; la donnée
        source contient des caractères actifs.
      operationId: suggest
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
          example: dolip
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        "200":
          description: Suggestions, médicaments et substances mêlés.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Suggestion" } }
                  meta: { $ref: "#/components/schemas/MetaListe" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/dataset:
    get:
      tags: [Exploitation]
      summary: État et provenance du jeu de données
      description: |
        Endpoint d'observabilité métier, et support de la conformité **Licence Ouverte** : il porte
        la mention de la source, de l'éditeur et de la date de mise à jour.
      operationId: getDataset
      responses:
        "200":
          description: Version, fraîcheur, compteurs, qualité d'ingestion et provenance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Dataset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NotReady" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Clé API d'au moins 32 caractères.
    adminAuth:
      type: http
      scheme: bearer
      description: Clé d'administration, distincte des clés API.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      description: |
        Curseur opaque de pagination. Il encode la position **et** la version du jeu de données :
        si celui-ci est rechargé entre deux pages, le curseur est rejeté avec `cursor_stale`.
        Mieux vaut une erreur explicite qu'une pagination silencieusement incohérente.
      schema: { type: string }
  headers:
    ETag:
      description: Empreinte de la réponse, à renvoyer dans `If-None-Match`.
      schema: { type: string }
    RequestId:
      description: Identifiant de la requête, à communiquer au support.
      schema: { type: string }
  responses:
    NotModified:
      description: La ressource n'a pas changé depuis l'`ETag` fourni.
...
  schemas:
...
    MetaSimple:
      type: object
      properties:
        dataset_version: { type: string }
...